Dried up create login in GraphQL
