+++ app/graphql/types/mutation_type.rb
@@ -3,11 +3,11 @@ module Types
     name 'Mutation'
     description 'Root query to mutate data'
 
-    field :createChallenge, function: ::Mutations::CreateChallenge.new
-    field :createCompletion, function: ::Mutations::CreateCompletion.new
-    field :createParticipation, function: ::Mutations::CreateParticipation.new
-    field :createRecordBook, function: ::Mutations::CreateRecordBook.new
-    field :createUser, function: ::Mutations::CreateUser.new
+    field :createChallenge, function: ::Mutations::Create::Challenge.new
+    field :createCompletion, function: ::Mutations::Create::Completion.new
+    field :createParticipation, function: ::Mutations::Create::Participation.new
+    field :createRecordBook, function: ::Mutations::Create::RecordBook.new
+    field :createUser, function: ::Mutations::Create::User.new
 
     field :login, function: ::Mutations::Login.new
